refactor: entrySet 적용 및 패키지 이동
