+++ src/main/java/chess/dto/StatusDto.java
@@ -50,8 +50,8 @@ public class StatusDto {
     public String toString() {
         StringBuilder sb = new StringBuilder();
         sb.append("{");
-        for (String color : score.keySet()) {
-            sb.append("\"").append(color).append("\" : ").append(score.get(color)).append(",");
+        for (Map.Entry<String, Double> entry : score.entrySet()) {
+            sb.append("\"").append(entry.getKey()).append("\" : ").append(entry.getValue()).append(",");
         }
         sb.append("\"").append("winner").append("\" : ").append("\"").append(winner).append("\"");
         sb.append("}");
